Move CLUSTER SETSLOT command to cluster.c
Move clusterParseSetSlotCommand and clusterCommandSetSlot from
cluster_legacy.c to cluster.c. The NODE subcommand now uses
clusterFinalizeSlotImport (when importing to myself) or
clusterProposeSlotChange (otherwise) via the vtable.

Replica migration after losing the last slot is handled in the
callback as a local decision, not a cluster-wide consensus operation.

Move setMigratingSlotDest, setImportingSlotSource, and
clusterCloseAllSlots from cluster_legacy.c to cluster_state.c as
they operate on common clusterState fields.

Expose clusterSetPrimary (was static in cluster_legacy.c) for use
by the SETSLOT callback.
