+++ src/cluster.h
@@ -169,6 +169,7 @@ void clusterDoBeforeSleep(int flags);
 void clusterSlotChange(slotRange *ranges, int numranges, clusterNode *target,
                        void *ctx, void (*callback)(void *ctx, int success));
 void clusterCleanupFailoverState(void);
+void clusterSetPrimary(clusterNode *n, int closeSlots, int full_sync_required);
 /* TODO: The following functions wrap legacy protocol details (CLUSTER_TODO_*
  * flags, manual failover timing). They should be replaced with protocol-agnostic
  * abstractions or moved behind the clusterBusType vtable when slot migration
